requirements/base.txt
coreapi==2.3.3
Django==2.0.1
djangorestframework==3.7.7
django-webpack-loader==0.5.0
Markdown==2.6.11
Pillow==5.0.0
Pygments==2.2.0
